2-URBAN_PLANNING_SYSTEM/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.1.0
async-timeout==4.0.3
attrs==21.2.0
Babel==2.8.0
backcall==0.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==4.1.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.6.4
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
elastic-transport==8.13.0
elasticsearch==8.13.0
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.13.4
fiona==1.9.6
fonttools==4.29.1
frozenlist==1.4.1
fs==2.4.12
fsspec==2024.3.1
gast==0.5.2
geopandas==0.14.4
gpg==1.16.0
greenlet==3.0.3
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.22.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
kiwisolver==1.3.2
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-text-splitters==0.0.1
langsmith==0.1.51
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
MarkupSafe==2.0.1
marshmallow==3.21.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
memory-profiler==0.61.0
more-itertools==8.10.0
mpmath==0.0.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.5.6
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
netifaces==0.11.0
networkx==3.3
notebook==6.4.8
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
orjson==3.10.1
osmnx==1.9.3
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prometheus-client==0.9.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproj==3.6.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
ranger-fm==1.9.3
regex==2024.4.28
reportlab==3.6.8
requests==2.31.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.1b0
shapely==2.0.4
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==2.0.29
sympy==1.9
systemd-python==234
tenacity==8.2.3
terminado==0.13.1
testpath==0.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
tornado==6.1
tqdm==4.66.2
traitlets==5.1.1
transformers==4.40.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==2.0.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zipp==1.0.0
